Random image index and add icons

+++ Raffle/Model/KLDrawBoxModel.h
@@ -11,7 +11,7 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-typedef NS_ENUM(NSUInteger, KLDrawMode) {
+typedef NS_ENUM(short, KLDrawMode) {
     KLDrawModeAttendee = 1,
     KLDrawModePrize = 2,
 };
